button: move button-png.c to button/
...so that all button code is in the same place.
