+++ src/button/meson.build
@@ -1,3 +1,4 @@
 labwc_sources += files(
+  'button-png.c',
   'button-xbm.c',
 )
